# docs/swagger.yaml
basePath: /
...
host: 192.168.55.42:8080
info:
  contact: {}
  description: API for an e-commerce platform with OTP-based authentication and product/market
    management.
  title: E-commerce API
  version: "1.0"
paths:
...
  /api/superadmin/banners:
    post:
      consumes:
      - multipart/form-data
      description: Creates a new banner with a description and thumbnail image. Requires
        superadmin JWT authentication.
      parameters:
      - description: Banner description (max 255 characters)
        in: formData
        name: description
        type: string
      - description: Thumbnail image (PNG/JPEG, max 5MB)
        in: formData
        name: thumbnail
        required: true
        type: file
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create a new banner
      tags:
      - Banners
...
  /api/superadmin/categories:
    post:
      consumes:
      - multipart/form-data
      description: Creates a category with name and optional thumbnail image. Requires
        superadmin JWT authentication.
      parameters:
      - description: Category name
        in: formData
        name: name
        required: true
        type: string
      - description: Category name_ru
        in: formData
        name: name_ru
        required: true
        type: string
      - description: Thumbnail image
        in: formData
        name: thumbnail
        type: file
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - Categories
...
  /api/superadmin/markets:
    post:
      consumes:
      - multipart/form-data
      description: Creates a new market with an admin account and optional thumbnail
        image. Requires superadmin JWT authentication.
      parameters:
      - description: Market name
        in: formData
        name: name
        required: true
        type: string
      - description: Market name russian
        in: formData
        name: name_ru
        required: true
        type: string
      - description: Market location
        in: formData
        name: location
        required: true
        type: string
      - description: Market location russian
        in: formData
        name: location_ru
        required: true
        type: string
      - description: Admin phone
        in: formData
        name: phone
        required: true
        type: string
      - description: Delivery price
        in: formData
        name: delivery_price
        required: true
        type: number
      - description: Admin password
        in: formData
        name: password
        required: true
        type: string
      - description: Thumbnail image
        in: formData
        name: thumbnail
        type: file
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create a new market
      tags:
      - Markets
...
securityDefinitions:
  BearerAuth:
    description: Bearer token for authentication (e.g., "Bearer <token>")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"